Former-commit-id: d552c96275e1d50684ae17ae43a17d9b96f66801

+++ app/controllers/baseListView.js
@@ -76,10 +76,14 @@ var handlers = (function(){
 				}
 			},
 			'destroy': function(model,col,options){
+				if( options && options.listItemDeletedAlready ){
+					return;
+				}
 				var len = $.section.getItems().length;
 				for(var i=0; i < len; i++){
 					if($.section.getItemAt(i).properties.itemId === model.get('id')){
 						$.section.deleteItemsAt(i, 1);
+						updateListMarker($.getCollection());
 						return;
 					}
 				}
@@ -123,8 +127,9 @@ var handlers = (function(){
 			'delete' : function(e) {
 				var item = $.getCollection().get(e.itemId);
 				item.destroy({
+					listItemDeletedAlready: true,
 					success: function(){
-						
+						updateListMarker($.getCollection());
 					},
 					error : function(){
 						alert('정상적으로 삭제하지 못했습니다. 새로고침 후 다시 시도해주세요.');
